feat: highlighting specific folders - uses url hash to auto scroll - saturate on hover and hash change
In a future commit the content immediately below the folder name will contain a clickable list of playlist names and folders. Other folders are on the same page, so clicking will highlight the folder on the page (now implemented)

This feature also allows any changes on the home page to be immediately highlighted to the user (i.e. they've saved a folder after editing or added a folder from another user's account).